@@ -2,7 +2,7 @@ namespace NoitaMap.Logging;
 
 public enum LogLevel
 {
-    Information,
-    Warning,
-    Critical
+    Info,
+    Warn,
+    Crit
 }
